refactor components to separate files
